chore: legend refactor
