+++ src/components/legend/Legend.js
