+++ src/components/legend/LegendItem.js
