+++ src/components/legend/LegendItemRange.js
