+++ src/components/legend/LineSymbol.js
