+++ src/components/legend/legendStyle.js
